chore: 거리 계산하는 함수 constant폻더내로 이동
